requirements.txt
pandas==0.22.0
python-telegram-bot==10.0.1
requests==2.18.4
bs4==0.0.1
geopy==1.12.0